requirements.txt
flask
flask-sqlalchemy
flask-wtf
flask-login
mysqlclient
python-dotenv